+++ electron/src/update/Squirrel.ts
@@ -24,8 +24,8 @@ import * as path from 'path';
 
 import * as config from '../js/config';
 import {getLogger} from '../js/getLogger';
-import * as lifecycle from '../js/lifecycle';
 import * as EnvironmentUtil from '../runtime/EnvironmentUtil';
+import * as lifecycle from '../runtime/lifecycle';
 
 app.setAppUserModelId(`com.squirrel.wire.${config.NAME.toLowerCase()}`);
